move link stat variables to 32bit to avoid any overflows

+++ src/main/rx/rx_unified_dsm.c
@@ -26,8 +26,8 @@ extern volatile uint8_t rx_frame_position;
 extern volatile uint8_t expected_frame_length;
 extern volatile frame_status_t frame_status;
 
-extern uint16_t link_quality_raw;
-extern uint8_t stat_frames_second;
+extern uint32_t link_quality_raw;
+extern uint32_t stat_frames_second;
 
 extern uint16_t bind_safety;
 extern int32_t channels[16];
